Arrumando Dijkstra

+++ api/src/main/java/Grafo/Caminho.java
@@ -4,7 +4,7 @@ import FaindMap.api.Vertice.Vertice;
 
 public class Caminho {
     public Vertice vertice = null;
-    public double estimativa;
+    public int estimativa;
     public int precedente;
     public Boolean aberto = true;
